Changing temperature and humidity data type to use decimal instead of int. Use datetime utc.

+++ MHIOT/Models/TemperatureData.cs
@@ -5,8 +5,8 @@ namespace MHIOT.Models
     public class TemperatureData
     {
         public int Id { get; set; }
-        public int? Temperature { get; set; }
-        public int? Humidity { get; set; }
+        public decimal? Temperature { get; set; }
+        public decimal? Humidity { get; set; }
         public DateTime? Date { get; set; }   
     }
 }
